test(cosmic-swingset): use access token
